[PERISTENCE/CORE]: slower but safer and more extensible 'framework'. (even though it's 1 Gb/s that I've lost on this commit :( ).

+++ persistence/raw_data.hpp
@@ -39,7 +39,7 @@ namespace neam
     /// \attention this class use the ownership memory management model, and you should be very cautionary when using this class.
     struct raw_data
     {
-      uint32_t size;
+      size_t size;
       int8_t *data;
 
       bool ownership;
